rename ai-client.ts
